+++ src/mvvm/mvvm.c
@@ -21,6 +21,8 @@
 #include "awtk.h"
 #include "mvvm/mvvm.h"
 #include "mvvm/base/mvvm_base.h"
+#include "mvvm/base/view_model_dummy.h"
+#include "mvvm/base/view_model_array_dummy.h"
 #include "mvvm/base/custom_binder.h"
 
 ret_t mvvm_init(void) {
@@ -29,6 +31,8 @@ ret_t mvvm_init(void) {
 
   log_debug("mvvm_init\n");
   return_value_if_fail(mvvm_awtk_init() == RET_OK, RET_FAIL);
+  view_model_factory_register(STR_VIEW_MODEL_DUMMY, view_model_dummy_create); 
+  view_model_factory_register(STR_VIEW_MODEL_ARRAY_DUMMY, view_model_array_dummy_create); 
 
   return RET_OK;
 }